refactor(query): enable little vacuum after query finish
