+++ src/query/service/src/interpreters/hook/mod.rs
@@ -14,6 +14,7 @@
 
 pub(crate) mod compact_hook;
 pub(crate) mod refresh_hook;
+pub(crate) mod vacuum_hook;
 
 #[allow(clippy::module_inception)]
 mod hook;
